merge master to duydat branch

+++ src/main/java/com/webphim/webphim/Reponsitory/ImageUserRepository.java
@@ -3,5 +3,5 @@ package com.webphim.webphim.Reponsitory;
 import com.webphim.webphim.Model.ImageUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ImageUserRepository extends JpaRepository<ImageUser, Integer> {
+public interface ImageUserRepository extends JpaRepository<ImageUser, Long> {
 }
